+++ src/biv_lite/app/plots.py
@@ -2,9 +2,8 @@ import typer
 from loguru import logger
 from pathlib import Path
 import pyvista as pv
-from biv_lite.biv_mesh import BivMesh
-from biv_lite.utils import to_pyvista_faces
-from biv_lite.vis import plot_biv_mesh
+from biv_lite import BivMesh
+from biv_lite.app.vis import to_pyvista_faces, plot_biv_mesh
 
 
 app = typer.Typer(help="Plot commands")