+++ src/biv_lite/meshing/utils.py
@@ -8,7 +8,3 @@ def flip_elements(mesh, material_id: int):
 
     return mesh
 
-# using pyvista format, you have to add number of points for each element
-def to_pyvista_faces(elements: np.ndarray) -> np.ndarray:
-    return np.hstack([np.ones((elements.shape[0], 1)) * 3, elements]).astype(np.int32)
-